backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
geopy==2.4.1
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4
reportlab==4.0.7
xlsxwriter==3.1.9
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3 